move makeDiscordCdnUrl to @cordis/util
