+++ libs/common/src/functions/index.ts
@@ -2,4 +2,3 @@ export * from './getMissingProps';
 export * from './halt';
 export * from './isPromise';
 export * from './keyMirror';
-export * from './makeDiscordCdnUrl';
