+++ libs/util/package.json
@@ -34,7 +34,8 @@
     "typescript": "^4.2.2"
   },
   "dependencies": {
-    "@cordis/rest": "workspace:^0.1.5",
+    "@cordis/rest": "workspace:^0.1.6",
+    "@cordis/error": "workspace:^0.1.6",
     "tslib": "^2.1.0"
   }
 }
