+++ libs/util/src/error.ts
