+++ libs/util/src/index.ts
@@ -1,4 +1,2 @@
-import type { RestManager } from '@cordis/rest';
-
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-export const makeRestUtils = (restRouter: RestManager) => ({});
+export * from './makeDiscordCdnUrl';
+export * from './restUtils';
